move test case to test_logger.cc
