+++ selfdrive/loggerd/logger.cc
@@ -1,5 +1,6 @@
 #include "selfdrive/loggerd/logger.h"
 
+#include <dirent.h>
 #include <sys/stat.h>
 #include <unistd.h>
 
@@ -123,6 +124,25 @@ static void lh_log_sentinel(LoggerHandle *h, SentinelType type) {
   auto bytes = msg.toBytes();
 
   lh_log(h, bytes.begin(), bytes.size(), true);
+}
+
+void clear_locks(const std::string &dir, const std::string &exclude_dir) {
+  DIR *d = opendir(dir.c_str());
+  if (!d) return;
+
+  struct dirent *entry;
+  std::string path;
+  while ((entry = readdir(d)) != nullptr) {
+    path = dir + "/" + entry->d_name;
+    if (entry->d_type == DT_DIR) {
+      if (strcmp(entry->d_name, ".") != 0 && strcmp(entry->d_name, "..") != 0 && path != exclude_dir) {
+        clear_locks(path, exclude_dir);
+      }
+    } else if (path.rfind(".lock") == (path.length() - 5)) {
+      unlink(path.c_str());
+    }
+  }
+  closedir(d);
 }
 
 // ***** logging functions *****
