+++ selfdrive/loggerd/tests/test_logger.cc
@@ -144,3 +144,27 @@ TEST_CASE("logger") {
     }
   }
 }
+
+TEST_CASE("clear_locks") {
+  const int cnt = 10;
+  for (int i = 0; i < cnt; ++i) {
+    std::string path = util::string_format("%s/%d", LOG_ROOT.c_str(), i);
+    REQUIRE(util::create_directories(path, 0775));
+    std::ofstream{path + "/.lock"};
+    REQUIRE(util::file_exists(path + "/.lock"));
+  }
+
+  clear_locks(LOG_ROOT, util::string_format("%s/%d", LOG_ROOT.c_str(), 0));
+
+  for (int i = 0; i < cnt; ++i) {
+    std::string path = util::string_format("%s/%d", LOG_ROOT.c_str(), i);
+    std::string lock_file = path + "/.lock";
+    if (i == 0) {
+      REQUIRE(util::file_exists(lock_file));
+      ::unlink(lock_file.c_str());
+    } else {
+      REQUIRE(util::file_exists(lock_file) == false);
+    }
+    rmdir(path.c_str());
+  }
+}
